+++ rpcalc/stack.py
@@ -24,9 +24,10 @@ class Stack:
     # note: these methods are ordered!
     # references to other methods are
     # given with a comment (ref:).
-    def __init__(self, initList, name):
+    def __init__(self, initList, name, limit=None):
         self.items = initList
         self.name = name
+        self.limit = limit
 
     def __getitem__(self, key):
         backKey = -1 * (key + 1)
@@ -38,6 +39,8 @@ class Stack:
     def __str__(self): #ref: len
         rep = self.name
         if self.name != '': rep += '\n'
+        if self.limit:
+            rep += "limited to " + str(self.limit) + " entries\n"
         backList = [ self[i] for i in range(len(self)) ]
         if len(self) != 0:
             longestEntry = max( [ len(str(i)) for i in self.items ] )
@@ -58,14 +61,35 @@ class Stack:
         backList = [ self[i] for i in range(len(self)) ]
         print(str(backList))
 
+    def clear(self):
+        if self.limit:
+            self.items = self.limit * [0.0]
+        else:
+            self.items = list()
+
     def push(self, item):
         self.items.append(item)
+        # If there is a stack size limit and the stack is overfull, erase the
+        # oldest entry
+        if self.limit:
+            if len(self) > self.limit:
+                self.items = self.items[1:]
 
     def pop(self):
-        #TODO define "is empty" (re: Alan)
+        #TODO define "is empty"
         if len(self) == 0:
-            print("empty stack!") #does this do anything?
-        else:
+            print("empty stack!") #FIXME does this do anything?
+        elif self.limit:
+            # With a limited stack, we duplicate the topmost (oldest) entry
+            # when the pop operation is called
+            toReturn = self.items.pop()
+            newList = [ self.items[0] ]
+            newList.extend(self.items)
+            self.items = newList
+            return toReturn
+        elif not self.limit:
+            # With an unlimited stack, we simply call the appropriate pop
+            # method on the underlying list
             return self.items.pop()
 
     def canOperate(self, argLen):