+++ changelog.md
@@ -3,6 +3,11 @@ Changelog
 
 ## __NEXT RELEASE__
 
+### Additions:
+- accepts command line arguments; try `rpcalc --help`
+- accepts a limited stack size (command line `-s`)
+- accepts initial values (command line `-i`)
+
 ## Version 0.6
 
 ### Fixes: